pc - clear the cache before doing a useSWR call on edit form
Otherwise, the data passed to the edit form may be stale data.
This is hacky, and it would be good to have a different solution,
since it violates the main idea of the "stale while revalidate"
pattern.  But for the time being, this is the best idea we
could come up with.